Added simple session tracking with play.Cache
After login, the user's username is stored in the session
then put into cache. Currently the main page prints out the
currently logged in user to test. This test first checks if the
current session has a uuid field, and if so, retrieves the username
from cache.

TODO: Create utility function, or maybe ControllerCache class
to deal with caching and sessions